dockerize

+++ backend/msvc-users/src/main/java/ec/edu/espe/msvc/users/services/UserServiceImpl.java
@@ -1,8 +1,11 @@
 package ec.edu.espe.msvc.users.services;
 
+import ec.edu.espe.msvc.users.client.CourseServiceClient;
+import ec.edu.espe.msvc.users.dto.EnrollmentDto;
 import ec.edu.espe.msvc.users.entity.User;
 import ec.edu.espe.msvc.users.dto.UserDto;
-import ec.edu.espe.msvc.users.repositories.UsuarioRepository;
+import ec.edu.espe.msvc.users.repositories.UserRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -14,7 +17,10 @@ import java.util.Optional;
 @Service
 public class UserServiceImpl implements UserService {
     @Autowired
-    private UsuarioRepository repository;
+    private UserRepository repository;
+
+    @Autowired
+    private CourseServiceClient courseServiceClient;
 
     @Override
     public List<User> getAll() {
@@ -38,9 +44,16 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
+    @Transactional
     public void delete(Long id) {
         if (!repository.existsById(id)) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "El id del usuario no existe o es incorrecto.");
+        }
+
+        List<EnrollmentDto> enrollments = courseServiceClient.getEnrollmentsByUserId(id);
+
+        if (!enrollments.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "El usuario tiene cursos registrados, por favor elimine las matrículas asociadas antes de eliminar el usuario.");
         }
 
         repository.deleteById(id);
@@ -73,6 +86,15 @@ public class UserServiceImpl implements UserService {
     @Override
     public List<UserDto> fetchUsersByIds(List<Long> ids) {
         return repository.findAllById(ids).stream().map(user -> UserDto.builder()
+                .id(user.getId())
+                .name(user.getName())
+                .email(user.getEmail())
+                .build()).toList();
+    }
+
+    @Override
+    public List<UserDto> search(String query) {
+        return repository.findByNameContainingIgnoreCase(query).stream().map(user -> UserDto.builder()
                 .id(user.getId())
                 .name(user.getName())
                 .email(user.getEmail())
